Move parsers to separate file, move empty file upload logic out of write_to_gcs
